scx_bpfland: convert scheduler crate into lib+bin crate
convert scheduler crate into lib+bin crate, to be able to
use(load/unload) scheduler by other crates

required for https://github.com/sched-ext/scx/pull/565
