+++ scheds/rust/scx_bpfland/Cargo.toml
@@ -25,3 +25,10 @@ scx_utils = { path = "../../../rust/scx_utils", version = "1.0.3" }
 
 [features]
 enable_backtrace = []
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "scx_bpfland"
+path = "src/main.rs"
